creating-custom-bevy-commands
Our `Snake` is moving but that's not showing on the screen just yet. This is because while we're spawning the `Entity` with the right `Position` in our `tick` function we don't have a `Sprite` component on that `Entity`.

The logic we need for spawning a `Snake` segment in `tick` is the same logic we used in `snake::spawn_snake`, so we have roughly two choices.

1. We could write a function that is not a system that we call from our systems
2. We could write a custom Bevy [`Command`](https://docs.rs/bevy/0.12.0/bevy/ecs/system/trait.Command.html)

For option 1, writing a function is fine but it requires that the arguments we pass to that function are queried in every system we use it in. We then have to pass all of those arguments into the function which is a bit of a headache for logic that uses a variety of arguments.

Using option 2, a custom `Command`, we'll be able to encapsulate all of the queries for `Resource`s and `Component`s where we're using them. The downside of using a custom `Command` is that we'll be dealing with the [`World`](https://docs.rs/bevy/0.12.0/bevy/ecs/prelude/struct.World.html) directly instead of getting system arguments injected for us.

Our `Command` will look like this when we use it to spawn a new segment, using [`commands.add`](https://docs.rs/bevy/latest/bevy/ecs/system/struct.Commands.html#method.add).

```rust
commands.add(SpawnSnakeSegment {
	position: Position::new(..)
});
```
\## Implementing a Command

In `src/snake.rs` we'll need to bring `Command` into scope.

```rust
use bevy::{ecs::system::Command, prelude::*};
```

and we can start implementing our new `Command`.

```rust
pub struct SpawnSnakeSegment {
    pub position: Position,
}

impl Command for SpawnSnakeSegment {
    fn apply(self, world: &mut World) {
	    todo!();
    }
}
```

`Command` is a trait, so that means we need a type to implement the trait for.

In `src/snake.rs`, we'll create a new struct called `SpawnSnakeSegment`. Any fields in this struct will be accessible when we apply this `Command`, so we'll store a `Position` that represents the location we want to spawn a new snake segment at.

To implement `Command` for `SpawnSnakeSegment`, we need to implement the `apply` function. `apply` takes `self`, which is a shorthand version of writing `self: Self`. `Self` is a way to refer to the type that's implementing the trait, which is a `SpawnSnakeSegment` in our case.

`apply` also gets an exclusive reference to `World`, which allows us to insert `Enitity`s, `Resource`, and more. The `World` type allows us to access pretty much anything in our Bevy app.

We've already implemented the logic for spawning in a new snake segment in `spawn_snake`, so we'll mostly copy that with two adjustments:

Instead of asking for `Board` from our system arguments, we'll use [`world.get_resource`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.World.html#method.get_resource) with a type argument.

```rust
let board = world.get_resource::<Board>().unwrap();
```

Instead of asking for `Snake` from our system arguments, we'll use [`world.get_resource_mut`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.World.html#method.get_resource_mut) with a type argument.

```rust
world.get_resource_mut::<Snake>().unwrap();
```

In the end we can remove almost all of the logic from our `spawn_snake` function and replace it with calls to our new `Command`. Our `src/snake.rs` file now looks like this.

```rust
use bevy::{ecs::system::Command, prelude::*};
use std::collections::VecDeque;

use crate::{
    board::{Board, Position, TILE_SIZE},
    colors,
};

\#[derive(Debug, Default, Resource)]
pub struct Snake {
    pub segments: VecDeque<Entity>,
}

pub fn spawn_snake(mut commands: Commands) {
    for position in
        [Position::new(3, 4), Position::new(4, 4)]
    {
        commands.add(SpawnSnakeSegment { position });
    }
}

pub struct SpawnSnakeSegment {
    pub position: Position,
}

impl Command for SpawnSnakeSegment {
    fn apply(self, world: &mut World) {
        let board = world.get_resource::<Board>().unwrap();
        let x = board
            .cell_position_to_physical(self.position.x);
        let y = board
            .cell_position_to_physical(self.position.y);

        let entity = world
            .spawn((
                SpriteBundle {
                    sprite: Sprite {
                        color: colors::SNAKE,
                        custom_size: Some(Vec2::splat(
                            TILE_SIZE,
                        )),
                        ..default()
                    },
                    transform: Transform::from_xyz(
                        x, y, 2.0,
                    ),
                    ..default()
                },
                self.position,
            ))
            .id();

        let mut snake =
            world.get_resource_mut::<Snake>().unwrap();

        snake.segments.push_front(entity);
    }
}
```

and in `src/lib.rs` in the `tick` system, we can replace our previous `commands.spawn(next_position)` with our new `Command`.

Bring `SpawnSnakeSegment` into scope.

```rust
use snake::{Snake, SpawnSnakeSegment};
```

and then use `commands.add` to apply our new `Command`.

```rust
pub fn tick(
    mut commands: Commands,
    mut snake: ResMut<Snake>,
    positions: Query<&Position>,
) {
    let snake_head_entity = snake
        .segments
        .front()
        .expect("snake should have a head entity");

    let next_position = positions.get(*snake_head_entity)
        .map(|head| {
            Position::new(head.x + 1, head.y)
        })
        .expect("stored entities in a snake should have a Position component associated with them");

    commands.add(SpawnSnakeSegment {
        position: next_position,
    });

    let old_tail = snake
        .segments
        .pop_back()
        .expect("snake should have a tail entity");
    commands.entity(old_tail).despawn_recursive();

    info!(?snake);
}
```

When we `cargo run` now, we see a snake moving to the right and off of the screen.

![snake running off screen](https://res.cloudinary.com/dilgcuzda/image/upload/v1700970112/workshops/snake-with-bevy-ecs/bevy-0.12/snake-off-screen.avif)

Let's finish up this lesson by removing the `info!(?snake);` log in `src/lib.rs`.