implementing-traits-for-newtypes
The snake is moving across the screen, but it's not in any way controlled by a user yet.

We could respond to user input keycodes in the `tick` system. Instead of `head + 1` we listen to the user's pressed keys and add or subtract to the relevant axis. We'll do this by implementing addition for our newtype.

In `tick` we'd add `Input<KeyCode>`

```rust
pub fn tick(
    mut commands: Commands,
    mut snake: ResMut<Snake>,
    positions: Query<&Position>,
    input: Res<Input<KeyCode>>,
) {
```

and then we can accomplish our goal by constructing a new `Position` in the direction we want to go, then adding that to the `head`'s `Position`.

```rust
let next_position = positions.get(*snake_head_entity)
    .map(|head| {
        let diff = if input.pressed(KeyCode::Up) {
            Position::new(0,1)
        } else if input.pressed(KeyCode::Down) {
            Position::new(0,-1)
        } else if input.pressed(KeyCode::Left) {
            Position::new(-1,0)
        } else if input.pressed(KeyCode::Right) {
            Position::new(1,0)
        } else {
            Position::new(0,0)
        };
        *head + diff
    })
    .expect("stored entities in a snake should have a Position component associated with them");
```

This has an error if you try to compile it, and that's that we can't add two `Position`s since we haven't implemented the [`Add`](https://doc.rust-lang.org/std/ops/trait.Add.html) trait for `Position`.

```rust
error[E0369]: cannot add `board::Position` to `board::Position`
  --> src/lib.rs:35:19
   |
35 |             *head + diff
   |             ----- ^ ---- board::Position
   |             |
   |             board::Position
   |
note: an implementation of `std::ops::Add` might be missing for `board::Position`
```

\## Implementing Add for newtypes

We have roughly two options here.

We can expose the inner `IVec2` in our `Position` and then add *those* together, before wrapping the `IVec2` in a `Position` again. This will expose the inner `IVec2` everywhere in our code because we'd need to make it `pub` for this to work.

```rust
pub struct Position(pub IVec2);
```

or we can implement `Add`, which is going to be the same logic but will allow us to use the `+` operator alone rather than wrapping/unwrapping our newtype.

> [!NOTE] derive_more
> If you find yourself implementing traits like Add for newtypes a lot, you may find the [derive_more](https://lib.rs/crates/derive_more) crate useful. This crate offers a bunch of extra derive macros to implement traits like Add, Mul, From, Display, and more. Often the implementation just uses whatever the underlying type's implementation is, which is exactly what we'll be doing here anyway.

In `src/board.rs`, right under where we defined `Position`, add an implementation of `Add` for `Position`.

```rust
\#[derive(Debug, Component, Deref, DerefMut, Clone)]
pub struct Position(IVec2);

impl Position {
    pub fn new(x: i32, y: i32) -> Self {
        Self(IVec2::new(x, y))
    }
}

impl Add for Position {
    type Output = Position;

    fn add(self, rhs: Self) -> Self::Output {
        Self(self.0 + rhs.0)
    }
}
```

The `Output` type is called an "associated type" and is the type that will be returned by our addition, which will be a new `Position`.

The `add` function takes the first `Position` as `self` and adds it to another `Position`: `rhs` (short for "right hand side").

The fully expanded type signature here makes it clear that we're using `Self` for both arguments, which is `Position` since that is the type we're implementing the trait for.

```rust
fn add(self: Self, rhs: Self) -> Self::Output
```

Since we're in the same module that we defined `Position` in, we can access private fields like the inner `IVec2`. This is what allows us to use `self.0` and `rhs.0`. We also need to wrap the resulting `IVec2` in a `Position`to return it.

```rust
Self(self.0 + rhs.0)
```

`Self` is the same as `Position` in this case.

With the trait implemented, we can `cargo run` once again and we see one more compilation error.
\## cannot move out of shared reference

That error is "cannot move out of `*head` which is behind a shared reference".

```
error[E0507]: cannot move out of `*head` which is behind a shared reference
  --> src/lib.rs:35:13
   |
35 |             *head + diff
   |             ^^^^^-------
   |             |
   |             `*head` moved due to usage in operator
   |             move occurs because `*head` has type `board::Position`, which does not implement the `Copy` trait
```

We implemented `Add` for `Position`, not `&Position`. Both are valid because references are first-class types and its common to implement one and use that implementation for the other.

By implementing `Add` for `Position` we made a decision about how `Position`. That decision is that we decided its cheap enough to copy the values in a `Position` when performing adds. This is *true* because primitive numbers work this way by default and what we have is basically two numbers. An `IVec2` after all is one `x` field and one `y` field.

However, we haven't told the Rust compiler that we consider it cheap to copy these values yet. We already have [`Clone`](https://doc.rust-lang.org/std/clone/trait.Clone.html) on `Position` which allows us to explicitly duplicate our struct by calling `.clone`. We can add [`Copy`](https://doc.rust-lang.org/std/marker/trait.Copy.html) to let the compiler do that for us implicitly when it needs to.

```rust
\#[derive(
    Debug, Component, Deref, DerefMut, Clone, Copy,
)]
pub struct Position(IVec2);
```

The technical term for this is "copy semantics", whereas normally Rust variable bindings have "move semantics". The big difference between "move" and "copy" is whether or not you can continue to use the original value afterwards. When "move"ing, you can't. When "copy"ing you can.

There is more on [Copy vs Clone](https://doc.rust-lang.org/std/marker/trait.Copy.html#whats-the-difference-between-copy-and-clone) in the docs.

\## Snake Movement

We can add two `Position`s now.

As far as our `Snake` movement logic goes: adding values in `tick` works, but it produces some odd behavior. The snake only moves when the user is pressing down a direction. We want to keep the snake moving in the same direction as the last pressed key when the user isn't pressing any keys.

In the next lesson, we'll make it so that the snake's movement continues in the user's desired direction even when the user isn't actively pressing a key.