keyboard-input-in-bevy
In the last lesson we implemented addition for `Position`s and allowed the user to control the movement of the snake, but the snake only moves when the user is pressing a key.

We want to keep the snake moving in the same direction as the last pressed key when the user isn't pressing any keys.

To do that we need to store the last pressed key somewhere. A `Resource` would make sense but there's one more glitch in the plan: the `tick` function only runs on the interval we set. That means you have to be actively pressing the key when the system runs to make the snake move.

To solve this we want to constantly be listening to user input, and update a resource with the last key the user pressed, then read the direction from _that_ resource any time the `tick` system runs.

This will allow the user to constantly give input while we read that input when we need it.

In `lib.rs` add a new submodule called `controls` and then create a new file called `controls.rs`.

```rust
pub mod board;
pub mod colors;
pub mod controls;
pub mod snake;
```

\## The controls Plugin

In `src/controls.rs` we'll define a new public enum named `Direction` to represent which direction the snake should be heading.

We're going to derive `Default` and `Resource` for `Direction`. When deriving `Default` on an enum, we have the opportunity to use the `#[default]` helper to pick which variant is the default variant, which will be `Right` since our snake will start off moving to the `Right`.

```rust
\#[derive(Resource, Default)]
pub enum Direction {
    Up,
    Down,
    Left,
    #[default]
    Right,
}
```

We'll also create a `user_input` system takes the user's input and a mutable reference to the `Direction` and set the `Direction`.

```rust
fn user_input(
    input: Res<Input<KeyCode>>,
    mut last_pressed: ResMut<Direction>,
) {
    if input.pressed(KeyCode::Up) {
        *last_pressed = Direction::Up;
    } else if input.pressed(KeyCode::Down) {
        *last_pressed = Direction::Down;
    } else if input.pressed(KeyCode::Left) {
        *last_pressed = Direction::Left;
    } else if input.pressed(KeyCode::Right) {
        *last_pressed = Direction::Right;
    }
}
```

We're using `Direction` instead of `KeyCode` here because we'll have to handle the current direction in `lib.rs` and using `Direction` communicates that there are indeed only four options, instead of the plethora of different keycodes we could act on.

If the relevant key is pressed, store that direction in the `Direction` resource.

Finally we need some way to expose this functionality to the rest of the application. We said we were going to be a bit more forward with our code organization in this app, so we'll bring in the Bevy prelude at the top of our file, and implement a `ControlsPlugin`.

Plugins are how, well, everything in Bevy is organized. You'll notice that if we don't add the `DefaultPlugins` in `main.rs`, then Bevy doesn't really do anything in the first place.

We have access to the same power for organizing our own code.

A Bevy `Plugin` allows us to take a mutable reference to an `App` and add resources and systems just like we would normally. So we do that for `Direction` and `user_input`.

Our plugin will initialize our `Direction` resource as well as adding the `user_input` system to the `Update` schedule.

```rust
use bevy::prelude::*;

pub struct ControlsPlugin;

impl Plugin for ControlsPlugin {
    fn build(&self, app: &mut App) {
        app.init_resource::<Direction>()
            .add_systems(Update, user_input);
    }
}
```

\## updating `tick`

In `tick`, we'll get the `controls::Direction` resource. Replace the `input` in the arguments of `tick` with `Res<controls::Direction`.

```rust
input: Res<controls::Direction>,
```

and we'll act on it like we did before using keycodes with a slight twist. We `match` on the value inside the `Resource` by dereferencing in the match.

```rust
let next_position = positions.get(*snake_head_entity)
    .map(|head| {
        let diff = match *input {
            controls::Direction::Up => Position::new(0,1),
            controls::Direction::Down => Position::new(0,-1),
            controls::Direction::Left => Position::new(-1,0),
            controls::Direction::Right => Position::new(1,0),
        };
        *head + diff
    })
    .expect("stored entities in a snake should have a Position component associated with them");
```

\## main.rs

Back in `main.rs`, we need to add our `ControlsPlugin` to our `App`, so bring the plugin into scope.

```rust
use snake::{
    board::{spawn_board, Board},
    controls::ControlsPlugin,
    snake::{spawn_snake, Snake},
    tick,
};
```

Then we'll add our plugin to the `add_plugins` call. Note that this needs to be a tuple of plugins, so you'll see extra parentheses at the beginning and end of the `.add_plugins` call.

```rust
.add_plugins((
    DefaultPlugins.set(WindowPlugin {
        primary_window: Some(Window {
            title: "Snake!".into(),
            ..default()
        }),
        ..default()
    }),
    ControlsPlugin,
))
```

And now we have our controls encapsulated in a custom plugin which we've added to our `App`, and the user can control the Snake as expectd.

```
cargo run
```