How about them apples?
A core component of Snake is the eating and spawning of apples.

In `lib.rs` create a public submodule called `food`.

```rust
pub mod board;
pub mod colors;
pub mod controls;
pub mod food;
pub mod snake;
```

In `food.rs` we'll create a struct that we'll use for tagging `Position`s as `Food` as well as a new `Command` to spawn apples. We've seen all of this before in this workshop so I'll show you the full code here.

```rust
use bevy::{ecs::system::Command, prelude::*};

use crate::{
    board::{Board, Position, TILE_SIZE},
    colors,
};

\#[derive(Component)]
pub struct Food;

pub struct SpawnApple {
    pub position: Position,
}

impl Command for SpawnApple {
    fn apply(self, world: &mut World) {
        let board = world.get_resource::<Board>().unwrap();
        let x = board
            .cell_position_to_physical(self.position.x);
        let y = board
            .cell_position_to_physical(self.position.y);

        world.spawn((
            SpriteBundle {
                sprite: Sprite {
                    custom_size: Some(Vec2::splat(
                        TILE_SIZE,
                    )),
                    color: colors::FOOD,
                    ..default()
                },
                transform: Transform::from_xyz(x, y, 2.0),
                ..default()
            },
            self.position,
            Food,
        ));
    }
}
```

Note that this logic is mostly a copy/paste of the snake segment command that adds an additional `Food` component.

We also need to add a color to our `colors` module for this to work.

In `src/colors.rs` add a color for `FOOD`. I've chosen `RED`.

```rust
pub const FOOD: Color = Color::RED;
```

\## Spawning the Apple

We don't really care about where we're spawning our first apple (both in terms of where it is in our code *and* in terms of the board position) because that will change later anyway. In this case we'll create a `spawn_apple` system in `src/food.rs` at position `(15,15)`.

```rust
pub fn spawn_apple(mut commands: Commands) {
    commands.add(SpawnApple {
        position: Position::new(15, 15),
    });
}
```

Then in `main.rs` bring `spawn_apple` into scope.

```rust
use snake::{
    board::{spawn_board, Board},
    controls::ControlsPlugin,
    food::spawn_apple,
    snake::{spawn_snake, Snake},
    tick,
};
```

and in our `Startup` schedule, append the `spawn_apple` system.

```rust
.add_systems(
    Startup,
    (
        setup,
        spawn_board,
        spawn_snake,
        spawn_apple,
    ),
)
```

Running `cargo run` now results in an apple on the screen!

\## Eating Apples

With an apple on the board, our snake needs to be able to actually eat them. When it eats an apple, it should grow into the apple space, making the snake one segment longer, and continue moving.

In `tick` in `lib.rs` add a query for `Position`s with `Food` components. [`With`](https://docs.rs/bevy/0.12.0/bevy/prelude/struct.With.html) is a query filter. The first tuple in the `Query` type arguments is the items we want to query for, while the second tuple is for filtering down our query to only the items we want.

In this case we want shared references to `Position` components and their associated `Entity`s that are labelled with a `Food` component.

Then at the bottom of our `tick` function, we'll replace our tail despawn logic.

```rust
let is_food = query_food
    .iter()
    .find(|(_, food_pos)| &&next_position == food_pos);
match is_food {
    Some((food_entity, _)) => {
        commands
            .entity(food_entity)
            .despawn_recursive();
    }
    None => {
        let old_tail = snake
            .segments
            .pop_back()
            .expect("snake should have a tail entity");
        commands.entity(old_tail).despawn_recursive();
    }
}
```

First we'll determine if the position our `Snake` is moving into next (our `next_position`) is a position in which a piece of `Food` exists. We do this by iterating over the `query_food` and using [`.find`](https://doc.rust-lang.org/std/iter/trait.Iterator.html#method.find) to test to see if the `next_position` is equal to the food position.

I chose to use an intermediate variable here called `is_food` for clarity. I do this whenever I feel like having the additional variable name clarifies what's happening in the code.

`find` returns us an `Option` so we can `match` on that and despawn the apple if our `find` found food, otherwise we'll despawn the tail as usual. This will extend the snake if it "eats" an apple and keep the snake the same length if it doesn't.