pausing-with-game-states
The Snake game is fairly playable now. It's recognizable as Snake in any case.

We still don't have any way to end the game or restart the game once it ends.

To do that we're going to introduce game states.

In `lib.rs` we can create a new enum to represent the states we could be in. We'll have one for the main menu and one for playing the game. Our `GameState` enum with derive [`States`](https://docs.rs/bevy/0.12.0/bevy/ecs/schedule/trait.States.html) as well as a number of other trait implementations.

```rust
\#[derive(
    Debug, Clone, Copy, Default, Eq, PartialEq, Hash, States,
)]
pub enum GameState {
    #[default]
    Menu,
    Playing,
}
```

We also need to add the state machine to our App in `src/main.rs`.

```rust
.add_state::<GameState>()
```

This game state creates a state machine that we can control via our Bevy systems. We can also use the state of the machine to control which systems are running at any given time.

In the `add_systems` call where we add the `tick` system we can use [`run_if`](https://docs.rs/bevy/0.12.0/bevy/ecs/prelude/trait.IntoSystemConfigs.html#method.run_if) to run systems conditionally if a [`Condition`](https://docs.rs/bevy/0.12.0/bevy/ecs/schedule/trait.Condition.html) is true.

In this case the `Condition` we're using is [`in_state`](https://docs.rs/bevy/0.12.0/bevy/ecs/prelude/fn.in_state.html) which will cause our system to run if the state machine is currently in the state we pass in. We've passed in  `GameState::Playing`.

```rust
.add_systems(
    FixedUpdate,
    tick.run_if(in_state(GameState::Playing)),
)
```

Now depending on which `GameState` variant is the `#[default]`, our game will either run as normal or will appear "paused" because our `tick` system is no longer running.